Rename openAI module and update usage
