+++ OpenAIHelper.py
